move all Space Object Representations in one file
