+++ src/ui/include/ui/UIFleet.hpp
@@ -9,7 +9,7 @@
 #include <ui_lib/Hover.hpp>
 #include <ui_lib/Line.hpp>
 #include <ui_lib/UIElement.hpp>
-#include <utils/RepresentationFleet.hpp>
+#include <utils/RepresentationSpaceObject.hpp.>
 
 
 namespace ui {
