+++ src/utils/include/utils/RepresentationSpaceObject.hpp
@@ -17,4 +17,18 @@ namespace utl {
         vec2pos_ty position{};
         SpaceObjectType type{};
     };
+
+    struct RepresentationFleet final : public RepresentationSpaceObject {
+        utl::usize engineProblemYears{};
+        RepresentationSpaceObject destRepresentation;
+    };
+
+    struct RepresentationPlanet final : public RepresentationSpaceObject {
+        bool isHumanPlayer{};
+        usize production{};
+    };
+
+    struct RepresentationTargetPoint final : public RepresentationSpaceObject { };
+
+    struct RepresentationBlackHole final : public RepresentationSpaceObject { };
 } // namespace utl
