--- src/utils/include/utils/RepresentationTargetPoint.hpp
@@ -1,12 +1,0 @@
-//
-// Purpur Tentakel
-// 17.01.2024
-//
-
-#pragma once
-
-#include "RepresentationSpaceObject.hpp"
-
-namespace utl {
-    struct RepresentationTargetPoint final : public RepresentationSpaceObject { };
-} // namespace utl
